Update ItemsService.java
Fix some errors

+++ src/main/java/com/csit321g2/Capstone/Service/ItemsService.java
@@ -55,7 +55,7 @@ public class ItemsService {
 		}
 	}
 	
-	public String deleteItem(int propertyTag) {
+	public String deleteItem(Long propertyTag) {
 		String msg = "";
 		
 		if (irepo.findById(propertyTag) != null) {
